services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
elevenlabs==1.9.0
pinecone-client==5.0.1
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
